Moved code from Bridge to Webpage.
The Album code in NodeBridge.js was a bit messy, so I moved it to JukeboxUI.html. This would also make it easier to modify for personal benefit.

As a side-effect, console logs in NodeBridge.js can't distinguish between whether it's sending a music disc or an album... not that it was important.